Use 7zip on windows
